% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{candidate.rxlr}
\alias{candidate.rxlr}
\title{Data files included in the package: candidate.rxlr is the result if the effector.summary function for new users that have not installed the additional software used by the effectR package}
\description{
Data files included in the package: candidate.rxlr is the result if the effector.summary function for new users that have not installed the additional software used by the effectR package
}
\author{
Javier F. Tabima \email{tabimaj@science.oregonstate.edu}
}
\keyword{data}
