% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_hmmer_path}
\alias{get_hmmer_path}
\title{Check if HMMER is installed and return path}
\usage{
get_hmmer_path(command, hmmer.path = NULL, error = TRUE, verbose = FALSE)
}
\arguments{
\item{command}{Which command to get the path for. For example "hmmsearch".}

\item{error}{If \code{TRUE}, throw an error if HMMER is not installed.}

\item{verbose}{If \code{TRUE}, print progress reports.}

\item{path}{Where to look for HMMER. By default it will look on the search
path.}
}
\value{
TRUE/FALSE
}
\description{
Check if HMMER is installed and return path to an executable
}
\keyword{internal}
