% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_mafft_path}
\alias{get_mafft_path}
\title{Check if MAFFT is installed and return path}
\usage{
get_mafft_path(mafft.path = NULL, error = TRUE, verbose = FALSE)
}
\arguments{
\item{error}{If \code{TRUE}, throw an error if MAFFT is not installed.}

\item{verbose}{If \code{TRUE}, print progress reports.}

\item{path}{Where to look for MAFFT. By default it will look on the search
path.}
}
\value{
TRUE/FALSE
}
\description{
Check if MAFFT is installed and return path to executable
}
\keyword{internal}
