% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.logo.R
\name{hmm.logo}
\alias{hmm.logo}
\title{Plots the relative frequencies of each position for \code{hmmsearch} table.}
\usage{
hmm.logo(hmm.table = hmm.table)
}
\arguments{
\item{hmm.table}{The HMM profile table resulting from \code{\link{hmm.search}}}
}
\description{
This function plots the results from \code{\link{hmm.search}} as a barplot with aminoacids in the x axis and the relative frequency of each aminoacid in the y axis
}
\keyword{effector}
\keyword{plot}
\keyword{regex}
