% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.search.R
\name{hmm.search}
\alias{hmm.search}
\title{Searching for motifs using HMM searches}
\usage{
hmm.search(original.seq = "file.fasta", regex.seq = sequences,
  mafft.path = "/usr/local/bin/", num.threads = 2,
  hmm.path = "/usr/local/bin/")
}
\arguments{
\item{original.seq}{The path for the original six-frame transaltion FASTA file}

\item{regex.seq}{A list of \code{SeqFastadna} objects resulting from \code{\link{regex.search}}. The HMM profile will be constructed using these sequences}

\item{mafft.path}{Local path of the MAFFT binary executable file}

\item{num.threads}{Number of threads to be used by MAFFT}

\item{hmm.path}{Local path of the HMMER binaries}
}
\value{
A list of three elements: REGEX candidate effectors, HMM candidate effectors, and HMM results table.
}
\description{
This function uses MAFFT and HMMER to search for sequences with RxLR or CRN motifs using hidden markov models.
}
\details{
\code{hmm.search} uses the results from \code{\link{regex.search}} to search for motifs of interest using hidden markov models after aligning the sequences with MAFFT.
After the multiple sequence alignment is complete, the function constructs a HMM profile using the alignment data. The HMM profile is in the original list of \code{SeqFastadna} objects to obtain the best HMM results with sequences with RxLR or CRN motifs.
}
\note{
The user has to specify the path for the MAFFT and the HMMER executable binaries and specify them in the \code{mafft.path} and \code{hmm.path}
}
\examples{

fasta.file <- system.file("extdata", "test_infestans.fasta", package = "effectR")
ORF <- seqinr::read.fasta(fasta.file)
REGEX <- regex.search(ORF, motif="RxLR")
\dontrun{
candidate.rxlr <- hmm.search(original.seq = fasta.file, regex.seq=REGEX, mafft.path="/usr/local/bin/", hmm.path="/usr/local/bin/", num.threads = 2)
}
}
\keyword{effector}
\keyword{regex}
