% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.search.R
\name{regex.search}
\alias{regex.search}
\title{Searching for motifs using regular expressions (REGEX)}
\usage{
regex.search(sequence, motif = "RxLR")
}
\arguments{
\item{sequence}{A list of \code{SeqFastadna} objects from \code{seqinr} \code{\link{read.fasta}}.}

\item{motif}{A character string indicating the motif to be searched (RxLR or CRN). Default is RxLR}
}
\description{
This function uses searches a list of \code{SeqFastadna} objects for sequences with RxLR or CRN motifs.
}
\examples{
fasta.file <- system.file("extdata", "test_infestans.fasta", package = "effectR")
ORF <- seqinr::read.fasta(fasta.file)
regex.search(ORF)

}
\keyword{effector}
\keyword{regex}
