% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{candidate.rxlr}
\alias{candidate.rxlr}
\title{An example of `effector.summary` output}
\description{
`candidate.rxlr` is an example result of the \code{\link{effector.summary}} function
for new users that have not installed the additional software used by the
effectR package.
}
\author{
Javier F. Tabima \email{tabimaj@science.oregonstate.edu}
}
\keyword{data}
