% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effector.summary.R
\name{effector.summary}
\alias{effector.summary}
\title{Returns non-redundant sequences from hmm.search or regex.search and generates a motif table}
\usage{
effector.summary(hmm.result, motif = "RxLR")
}
\arguments{
\item{hmm.result}{A list of \code{SeqFastadna} objects obtained from \code{\link{regex.search}} or \code{\link{hmm.search}}}

\item{motif}{A character string indicating the motif to be searched (RxLR or CRN). Default is RxLR}
}
\value{
A list of two objects: Summary motif table and non-redundant sequences (only with results of \code{\link{hmm.search}})
}
\description{
This function summarize the results from \code{\link{regex.search}} or \code{\link{hmm.search}}.
}
\examples{
\dontrun{

fasta.file <- system.file("extdata", "test_infestans.fasta", package = "effectR")
ORF <- seqinr::read.fasta(fasta.file)
REGEX <- regex.search(ORF, motif='RxLR')
candidate.rxlr <- hmm.search(original.seq = fasta.file, regex.seq=REGEX, num.threads = 2)
effector.summary(candidate.rxlr)
}
}
\keyword{effector}
\keyword{regex}
