% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.effectR.R
\name{shiny.effectR}
\alias{shiny.effectR}
\title{Function to run the effectR package using a shiny app}
\usage{
shiny.effectR()
}
\description{
This function will launch an interactive web interface to run the effectR package functions to obtain effectors. It runs using the \pkg{shiny} R package.
}
\details{
To sucessfully run this function the user will need to set the
}
