% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.effectR.R
\name{shiny.effectR}
\alias{shiny.effectR}
\title{Function to run the effectR package using a shiny app}
\usage{
shiny.effectR(mafft.path = NULL, hmm.path = NULL)
}
\arguments{
\item{mafft.path}{Local path of folder containing the MAFFT binary executable file or the executable file itself. If not specified, then MAFFT must be in the program search path.}

\item{hmm.path}{Local path of  folder containing the HMMER binaries.  If not specified, then HMMER executables must be in the program search path.}
}
\description{
This function will launch an interactive web interface to run the effectR package functions to obtain effectors. It runs using the \pkg{shiny} R package.
}
\details{
To successfully run this function the user will need to set the
}
